Restructured module directory structure to make more sense
